refactor(sift): hoist the descriptor's batch helpers, stop re-zeroing scratch
Cleanup pass over this branch's optimisation work.

* `exp_batch` / `mag_ang_batch` move from `descriptor.rs` into `hal.rs`, beside
  the `x4` primitives they drive.
* `DescriptorScratch` no longer clears and re-`resize`s its `mag`/`ang` buffers
  per keypoint. Both are fully overwritten before they are read, so the zero
  fill never survived — it was writing ~1400 floats per keypoint for nothing.
  `grow_to` now grows them monotonically instead. `clear()` documents why those
  two fields are deliberately not cleared.

Deliberately NOT shared: `orient.rs` expresses the same four-lane dispatch and
was routed through the new helpers, then reverted. Measured back to back,
orientation went 19.0 -> 22.5 ms, an 18% regression — at a fixed width of four
the helpers' loop preambles do not fold away, and `#[inline]` did not recover
it, whereas the straight-line block has none. The duplication is the price of
that and the comment in `orient.rs` records it so the next cleanup pass does
not silently pay it again.

Also fixes a build break introduced while hoisting: inserting the helpers
directly above `pub mod x4` stole that module's
`#[cfg(target_arch = "aarch64")]`, leaving `x4`'s `use std::arch::aarch64::*`
ungated so the crate did not compile off aarch64 at all. This is the
scripted-insert attribute-theft trap; `cargo clippy --target
x86_64-unknown-linux-gnu` is what catches it and is now clean, along with the
aarch64 clippy and a featureless `cargo check -p kornia-py`.

54/54 bitwise oracle tests pass across all ten images; single-thread frame time
unchanged at ~207 ms.
